+++ addons/web/__manifest__.py
@@ -203,9 +203,6 @@ This module provides the core of the Odoo Web Client.
             'web/static/src/env.js',
             'web/static/src/core/utils/transitions.scss', # included early because used by other files
             'web/static/src/core/**/*',
-            # TODO: split debug files correctly when converting frontend debug
-            # manager so we don't need action-service and stuff
-            ('remove', 'web/static/src/core/debug/**/*'),
             'web/static/src/public/error_notifications.js',
 
             'web/static/src/legacy/scss/base_frontend.scss',
