+++ addons/web/static/src/core/debug/debug_context.js
@@ -60,8 +60,12 @@ class DebugContext {
 }
 
 const debugContextSymbol = Symbol("debugContext");
+export function createDebugContext(env, { categories = [] } = {}) {
+    return { [debugContextSymbol]: new DebugContext(env, categories) };
+}
+
 export function useOwnDebugContext({ categories = [] } = {}) {
-    useSubEnv({ [debugContextSymbol]: new DebugContext(useEnv(), categories) });
+    useSubEnv(createDebugContext(useEnv(), { categories }));
 }
 
 export function useCurrentDebugContext() {
