--- addons/web/static/src/legacy/frontend/debug_manager/debug.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-
-<t t-name="WebClient.DebugManager">
-    <li class="o_debug_manager" role="menuitem">
-        <t t-set="_devtool_button_title">Open Developer Tools</t>
-        <a role="button" href="#" class="o_debug_mode"
-                         t-att-title="_devtool_button_title + widget.debug_mode_help"
-                         t-att-aria-label="_devtool_button_title + widget.debug_mode_help"
-                         t-att-data-debug-mode="widget.debug_mode"
-                         data-toggle="dropdown" aria-expanded="false" tabindex="-1" data-display="static">
-            <span class="fa fa-bug">
-                <span class="profiling_items text-danger fa fa-circle"></span>
-            </span>
-        </a>
-        <div class="dropdown-menu dropdown-menu-right o_debug_dropdown" role="menu"/>
-    </li>
-</t>
-<t t-name="WebClient.DebugManager.Global">
-    <a role="menuitem" href="#" data-action="split_assets" class="dropdown-item o_debug_split_assets">Activate Assets Debugging</a>
-    <a role="menuitem" href="#" data-action="tests_assets" class="dropdown-item">Activate Tests Assets Debugging</a>
-    <a role="menuitem" href="#" data-action="regenerateAssets" class="dropdown-item">Regenerate Assets Bundles</a>
-    <a t-if="manager._is_admin" role="menuitem" href="/web/become" class="dropdown-item">Become Superuser</a>
-    <a role="menuitem" href="#" data-action="leave_debug_mode" class="dropdown-item">Leave the Developer Tools</a>
-    <div class="dropdown-divider" role="separator"/>
-    <span role="menuitem" class="dropdown-item profiling">
-        <span class="custom-control custom-switch">
-            <input type="checkbox" class="custom-control-input" id="enable_profiling"/>
-            <label class="custom-control-label" for="enable_profiling">
-                Enable profiling
-                <span class="profiling_items text-danger fa fa-circle"></span>
-            </label>
-            <i id="open_profiling_button" class="open_profiling float-right fa fa-list fa-lg"></i>
-        </span>
-    </span>
-    <span class="profiling_items">
-        <span role="menuitem" class="dropdown-item profiling">
-            <span class="custom-control custom-switch">
-                <input type="checkbox" class="custom-control-input profile_switch" id="profile_sql"/>
-                <label class="custom-control-label" for="profile_sql">Record sql</label>
-            </span>
-        </span>
-        <span role="menuitem" class="dropdown-item profiling">
-            <span class="custom-control custom-switch">
-                <input type="checkbox" class="custom-control-input profile_switch" id="profile_traces_async"/>
-                <label class="custom-control-label" for="profile_traces_async">Record traces</label>
-            </span>
-        </span>
-        <span role="menuitem" class="dropdown-item profiling profile_traces_async">
-            <span class="">
-                <div class="input-group input-group-sm">
-                    <div class="input-group-prepend">
-                        <div class="input-group-text">Interval</div>
-                    </div>
-                    <select id="profile_traces_async_interval" class="profile_param form-control">
-                        <option value="">Default</option>
-                        <option value="0.001">0.001</option>
-                        <option value="0.01">0.01</option>
-                        <option value="0.1">0.1</option>
-                        <option value="1">1</option>
-                    </select>
-                    <!--input type="number" step="any" min="0.001" max="10" value="0.001" /-->
-                </div>
-            </span>
-        </span>
-    </span>
-</t>
-
-</templates>
